Add support for RegExp keys
If a rel is not an IANA rel, it should be a FQDN. This means that there could be some variance in the rel string, depending on where said rels are defined. This change supports passing a regex to the various `hasXByY` and `getXByY` functions, which allows for cases where the entire rel can vary to be handled cleanly.

(The other option is to basically offload this work outside the client, and not use the various helper methods.)