Fix "entity is undefined"
This was noticed in d2l-my-courses, which uses d2l-siren-parser, which uses this. If the passed entity is undefined for whatever reason, then things crap out on trying to find out if the rel exists. To safely navigate this, and always ensure entity.getXByY and other functions work, return an empty entity if undefined or null was passed.