Improve error messages to include problem
It has happened a few times now that a Siren object defined in a test is missing a required attribute, or it is of the wrong type, but without specifying more information, it was often difficult to track down these mistakes. This adds a little more information, although in the case where something is missing it just says "got undefined", which is perhaps of dubious value but at least it makes it abundantly clear what exactly is wrong.