Homogenize two versions
There have been som issues that arise from using the non-global version of the siren parser. Since there's no downside to using the -global version in contexts were it's not required (but there is a downside to using the non-global one where global one is required), we might as well just have one, deumdified, version.

For this (minor) release, we are simply making the two files the same. In the next major release, the -global.js file will be removed, as the new siren-parser.js is exactly equivalent to the old siren-parser-global.js.