US124690 Update and modernize
- Update dependencies
- Add package-lock.json
- Switch from Travis CI to GitHub Actions

This removes the `report-cov` command from the package.json, as it
hasn't been working anyway. (There is no Coveralls token in this repo
either, so not sure if it ever did?) The dependency updates here are all
either minor/patch versions, or are depedencies only used in tests
anyway. Beyond that, the only other changes are small changes required
by the new eslint rules.