Undefined should return false
Due to a long-standing error in how `contains` and `getMatchingValue`
work, this library was returning the first X for any `hasX/getX` call if
undefined was passed. This is because the check we ended up doing was
`'somestring'.match(undefined)`, which returns an array, which is
truthy, meaning we'd just return the first thing. This adds a check for
falsey values, so that we don't return anything for undefined/null.