Update readme with new/correct import
tl;dr [this StackOverflow post](https://stackoverflow.com/a/33705077/271473).

Turns out Babel 5 had a hack in place for default exports such that when there was _only_ a default export, it was set as just `module.exports`. However, with the new exports introduced in #39, that hack no longer applies, so our export is now:

```js
{
  default: Entity,
  Link,
  Action,
  Field
}
```

This means that existing usages need to be updated slightly - arguably this could be seen as a breaking change, but in reality this is more akin to fixing a weird behaviour that has been sitting in wait since we updated to Babel >5.